reorganize estimator inheritence structure
