--- ff_estimate/include/ff_estimate/base_estimator.hpp
@@ -1,56 +1,0 @@
-// MIT License
-//
-// Copyright (c) 2023 Stanford Autonomous Systems Lab
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
-
-
-#pragma once
-
-#include <string>
-
-#include <rclcpp/rclcpp.hpp>
-
-#include "ff_msgs/msg/free_flyer_state.hpp"
-#include "ff_msgs/msg/free_flyer_state_stamped.hpp"
-
-namespace ff
-{
-
-/**
- * @brief estimator base class
- */
-class BaseEstimator : public rclcpp::Node
-{
-public:
-  explicit BaseEstimator(const std::string & node_name);
-
-protected:
-  /**
-   * @brief send out the current state estimate
-   *
-   * @param state estimated freeflyer state
-   */
-  void SendStateEstimate(const ff_msgs::msg::FreeFlyerState & state);
-
-private:
-  rclcpp::Publisher<ff_msgs::msg::FreeFlyerStateStamped>::SharedPtr state_pub_;
-};
-
-}  // namespace ff
